refactor: add initRandomColorButton
