+++ frontend/src/js/login_panel.js
@@ -4,7 +4,7 @@ import { addEventListenerTo } from "./ui_utils/ui_utils.js";
 import { getColor, getRandomColor } from "./ui_utils/color_utils.js";
 import { resetInputField, setInputFieldHint } from "./ui_utils/input_field_utils.js";
 
-import { initBackButton } from "./ui_utils/button_utils.js"
+import { initBackButton, initRandomColorButton } from "./ui_utils/button_utils.js"
 import { initTogglePasswordVisibilityIcon } from "./ui_utils/input_field_utils.js";
 import { loadMainMenuPanel } from "./main_menu_panel.js";
 import { postRequest } from "./network_utils/api_requests.js";
@@ -17,65 +17,18 @@ export async function loadLoginPanel() {
     await loadComponent('components/login_panel.html')
 
     initBackButton(() => loadStartPanel())
-    initConfirmLoginButton()
+    initRandomColorButton(
+      'confirm-login-button',
+      'login-panel',
+      () => {
+        // handleLogin()
+        load2FAPanel()
+      }
+    )
     initTogglePasswordVisibilityIcon()
   } catch (error) {
     console.error('Error loading Login Panel:', error)
   }
-}
-
-function initConfirmLoginButton() {
-  const button = document.getElementById('confirm-login-button')
-  let colorInfo = {
-    hex: '',
-    name: ''
-  }
-
-  addEventListenerTo(
-    button,
-    'click',
-    () => {
-      // handleLogin()
-      load2FAPanel()
-    }
-  )
-
-  addEventListenerTo(
-    button,
-    'mouseover',
-    () => {
-      colorInfo = getRandomColor(500)
-
-      button.style.backgroundColor = colorInfo['hex']
-      button.style.color = getColor(colorInfo['name'], 800)
-    }
-  )
-
-  addEventListenerTo(
-    button,
-    'mouseout',
-    () => {
-      button.style.backgroundColor = getColor('charcoal', 700)
-      button.style.color = getColor('charcoal', 100)
-    }
-  )
-
-  addEventListenerTo(
-    button,
-    'mousedown',
-    () => {
-      button.style.backgroundColor = getColor(colorInfo['name'], 700)
-    }
-  )
-
-  addEventListenerTo(
-    button,
-    'mouseup',
-    () => {
-      button.style.backgroundColor = colorInfo['hex']
-      button.style.color = getColor(colorInfo['name'], 800)
-    }
-  )
 }
 
 async function handleLogin() {
